BACKEND - AUTENTICAÇÃO BÁSICA

+++ src/main/java/io/horacerta/repository/PontoDao.java
@@ -1,14 +1,13 @@
 package io.horacerta.repository;
 
 
-import io.horacerta.model.Pessoa;
-import io.horacerta.model.PontoDiario;
-
 import java.util.Date;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface PontoDao extends CrudRepository<PontoDiario, Integer>,PontoDaoCustom{
+import io.horacerta.model.PontoDiario;
+
+public interface PontoDao extends CrudRepository<PontoDiario, Long>,PontoDaoCustom{
 	
 //   PontoDiario findByDataRegistroAndPessoa(Date dataRegistro, Integer idPessoa);
    PontoDiario findByDataRegistro(Date dataRegistro);
